Make keybindings configurable
Make ner have keybindings that are configurable in the .ner.yaml
configuration file.  The keybindings can be set in different sections
in the configuration file, and missing entries, or sections will make
ner use the default keybindings.

The sample configuration file is updated with the default values for
the keys.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>